basic itts layout and routing
