+++ frontend/src/Footer.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+ 
+class Footer extends Component {
+  render() {
+    return (
+      <footer>
+          FOOTER
+      </footer>
+    );
+  }
+}
+
+export default Footer;
