+++ frontend/src/Header.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+ 
+class Header extends Component {
+  render() {
+    return (
+      <header>
+          IM TELLING THE STORY
+      </header>
+    );
+  }
+}
+
+export default Header;
