+++ frontend/src/Main.js
@@ -1,31 +1,25 @@
 import React, { Component } from "react";
 import {
-  Route,
-  NavLink,
-  HashRouter
-} from "react-router-dom";
-import Home from "./Home";
-import Stuff from "./Stuff";
-import Contact from "./Contact";
+    HashRouter
+  } from "react-router-dom";
+import Header from "./Header";
+import Content from "./Content";
+import Footer from "./Footer";
+import ProgressBar from "./ProgressBar";
+import PrimaryNav from "./PrimaryNav";
 
 class Main extends Component {
   render() {
-    return (
+    return (        
         <HashRouter>
-        <div>
-          <h1>Simple SPA</h1>
-          <ul className="header">
-            <li><NavLink exact to="/">Home</NavLink></li>
-            <li><NavLink to="/stuff">Stuff</NavLink></li>
-            <li><NavLink to="/contact">Contact</NavLink></li>
-          </ul>
-          <div className="content">
-            <Route exact path="/" component={Home}/>
-            <Route path="/stuff" component={Stuff}/>
-            <Route path="/contact" component={Contact}/>
-          </div>
-        </div>
-      </HashRouter>
+            <div class="container">
+                <Header></Header>
+                <ProgressBar></ProgressBar>
+                <Content></Content>
+                <PrimaryNav></PrimaryNav>
+                <Footer></Footer>
+            </div>
+        </HashRouter>
     );
   }
 }
