+++ frontend/src/ProgressBar.js
@@ -1,0 +1,13 @@
+import React, { Component } from "react";
+ 
+class ProgressBar extends Component {
+  render() {
+    return (
+      <div class="progress-bar">
+          PROGRESS BAR
+      </div>
+    );
+  }
+}
+
+export default ProgressBar;
