+++ frontend/src/Vote.js
@@ -1,10 +1,10 @@
 import React, { Component } from "react";
  
-class Contact extends Component {
+class Vote extends Component {
   render() {
     return (
       <div>
-        <h2>GOT QUESTIONS?</h2>
+        <h2>Vote</h2>
         <p>The easiest thing to do is post on
         our <a href="http://forum.kirupa.com">forums</a>.
         </p>
@@ -13,4 +13,4 @@ class Contact extends Component {
   }
 }
  
-export default Contact;
+export default Vote;
